Bug 938680 - Don't show "Remove X" button for yourself
When a curator is viewing a group, the group members are shown
withe a "Remove X" button that lets the curator remove the user
from the group. Curators should not be able to remove themselves
from the groups they curate, though. So don't show a "Remove X"
button on a user's own member card when the user is a curator
viewing a group they curate.

Fix bug 938680